requirements.txt
pandas
mlflow
uvicorn[standard]
fastapi
lightfm
xgboost
sentence-transformers
faiss-cpu
scikit-learn
streamlit
pyarrow